ruby-2.2.0, rails-4.2.0
