+++ .travis.yml
@@ -1,6 +1,6 @@
 language: ruby
 rvm:
-  - 2.1.2
+  - 2.2.0
 
 services:
   - mongodb
