+++ config/environments/test.rb
@@ -8,7 +8,7 @@ Rails.application.configure do
   config.action_controller.include_all_helpers = false
 
   # Configure static asset server for tests with Cache-Control for performance.
-  config.serve_static_assets  = true
+  #config.serve_static_assets  = true
   #config.static_cache_control = "public, max-age=3600"
 
   # Show full error reports and disable caching.
