+++ spec/spec_helper.rb
@@ -7,7 +7,7 @@ ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require File.expand_path(__FILE__, "../../app/helpers")
 require 'rspec/rails'
-require 'rspec/autorun'
+#require 'rspec/autorun'
 require 'capybara/rspec'
 require 'capybara/rails'
 
@@ -26,45 +26,6 @@ SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
 SimpleCov.start do
   add_filter 'spec/'
   add_filter 'vendor/bundle'
-end
-
-def can_test_mecab_spec?
-  # In Travis Ci dictionares_controller#build failed because of there is no MeCab installed.
-  # So, before test, check whether we can do MeCab specific tests.
-  return false if SS.config.kana.disable
-  unless ::File.exists?(SS.config.kana.mecab_indexer)
-    puts("[MeCab Spec] not found: #{SS.config.kana.mecab_indexer}")
-    return false
-  end
-  unless ::File.exists?(SS.config.kana.mecab_dicdir)
-    puts("[MeCab Spec] not found: #{SS.config.kana.mecab_dicdir}")
-    return false
-  end
-  true
-end
-
-def can_test_open_jtalk_spec?
-  # be carefule, open jtalk spec is slow.
-  # you will waste a lot of time if you turn on allow_open_jtalk.
-  return false if ENV["allow_open_jtalk"].to_i == 0
-  return false if SS.config.voice.disable
-  unless ::File.exists?(SS.config.voice['openjtalk']['bin'])
-    puts("[Open JTalk Spec] not found: #{SS.config.voice['openjtalk']['bin']}")
-    return false
-  end
-  unless ::Dir.exists?(SS.config.voice['openjtalk']['dic'])
-    puts("[Open JTalk Spec] not found: #{SS.config.voice['openjtalk']['dic']}")
-    return false
-  end
-  unless ::File.exists?(SS.config.voice['openjtalk']['sox'])
-    puts("[Open JTalk Spec] not found: #{SS.config.voice['openjtalk']['sox']}")
-    return false
-  end
-  unless ::File.exists?(SS.config.voice['lame']['bin'])
-    puts("[Open JTalk Spec] not found: #{SS.config.voice['lame']['bin']}")
-    return false
-  end
-  true
 end
 
 RSpec.configure do |config|
@@ -95,6 +56,9 @@ RSpec.configure do |config|
   #     --seed 1234
   #config.order = "random"
   config.order = "order"
+
+  config.include Rails.application.routes.url_helpers
+  config.include Capybara::DSL
 
   config.include FactoryGirl::Syntax::Methods
   config.before do
